Simplify Command API
Instead of extending from a base abstract class, which wasn't
passing the generic types through the methods the way I'd like,
provide `Command` as an interface the command object has
to conform to. Since we're loading up these commands
dynamically, we also need to validate them, so we use io-ts
to generate a validator for this. This is much easier to implement
and get the types right on the implementors side vs the manual
typing needed when implementing against an abstract class.
It also means we no longer have to import from `brookjs-cli`,
which could cause module conflicts if we don't get the same
instance of the `Command` constructor.

All of the commands are refactored over to ensure they line
up with the new interface, as well as the tests. Use `useDeltas`
instead of the custom `Root` component we wrote previously.
Update the tests to use Jest's mocks instead of DI. Some logic
was extracted to separate modules for code cleanliness.
