+++ packages/brookjs-cli/features/commands.feature
@@ -6,31 +6,21 @@ Feature: cli commands
 
   @valid
   @js
-  Scenario: Developer adds a valid command
+  Scenario: Developer adds a valid js command
     Given I have a project
     And I add a "js" command with body
       """
       import React from 'react';
       import Kefir from 'kefir';
-      import { Command } from 'brookjs-cli';
       import { Text } from 'ink';
 
-      export class ValidCommand extends Command {
+      export const ValidCommand = {
         builder(yargs) {
           return yargs;
-        }
-
-        cmd = 'valid';
-
-        describe = 'A valid command!';
-
-        initialState = () => ({});
-
-        exec = () => () => Kefir.never();
-
-        reducer = state => state;
-
-        View = () => <Text>Success from ValidCommand!</Text>;
+        },
+        cmd: 'valid',
+        describe: 'A valid command!',
+        View: () => <Text>Success from ValidCommand!</Text>,
       }
       """
     When I run beaver with "valid"
@@ -40,9 +30,34 @@ Feature: cli commands
       Success from ValidCommand!
       """
 
+    @invalid
+    @js
+    Scenario: Developer adds an invalid js command
+      Given I have a project
+      And I add a "js" command with body
+        """
+        import React from 'react';
+        import Kefir from 'kefir';
+        import { Text } from 'ink';
+
+        export const InvalidCommand = {
+          builder(yargs) {
+            return yargs;
+          },
+          cmd: 'valid',
+          View: () => <Text>Success from ValidCommand!</Text>,
+        }
+        """
+      When I run beaver with "valid"
+      And I wait for the command to finish with code 0
+      Then I see this in stdout
+        """
+        InvalidCommand is not a valid command.
+        """
+
     @valid
     @ts
-    Scenario: Developer adds a valid command
+    Scenario: Developer adds a valid ts command
       Given I have a project
       And I add a "tsx" command with body
         """
@@ -51,22 +66,13 @@ Feature: cli commands
         import { Command } from 'brookjs-cli';
         import { Text } from 'ink';
 
-        export class ValidCommand extends Command<any, any, any, any> {
+        export const ValidCommand: Command<{}> = {
           builder(yargs) {
             return yargs;
-          }
-
-          cmd = 'valid';
-
-          describe = 'A valid command!';
-
-          initialState = () => ({});
-
-          exec = () => () => Kefir.never();
-
-          reducer = state => state;
-
-          View = () => <Text>Success from ValidCommand!</Text>;
+          },
+          cmd: 'valid',
+          describe: 'A valid command!',
+          View: () => <Text>Success from ValidCommand!</Text>,
         }
         """
       When I run beaver with "valid"
@@ -75,3 +81,29 @@ Feature: cli commands
         """
         Success from ValidCommand!
         """
+
+      @invalid
+      @ts
+      Scenario: Developer adds an invalid ts command
+        Given I have a project
+        And I add a "tsx" command with body
+          """
+          import React from 'react';
+          import Kefir from 'kefir';
+          import { Command } from 'brookjs-cli';
+          import { Text } from 'ink';
+
+          export const InvalidCommand: Command<{}> = {
+            builder(yargs) {
+              return yargs;
+            },
+            cmd: 'valid',
+            View: () => <Text>Success from ValidCommand!</Text>,
+          }
+          """
+        When I run beaver with "valid"
+        And I wait for the command to finish with code 0
+        Then I see this in stdout
+          """
+          InvalidCommand is not a valid command.
+          """
