+++ packages/brookjs-cli/src/commands/BuildCommand/exec.ts
@@ -1,10 +1,11 @@
 import Kefir, { Stream, Property } from 'kefir';
 import { Nullable } from 'typescript-nullable';
+import { WebpackService } from '../../services';
 import { webpackBuild } from './actions';
 import { selectWebpackConfig } from './selectors';
 import { State, Action } from './types';
 
-const exec = ({ WebpackService }: typeof import('../../services')) => (
+const exec = (
   action$: Stream<Action, never>,
   state$: Property<State, never>
 ): Stream<Action, never> =>
