+++ packages/brookjs-cli/src/commands/TestCommand/initialState.ts
@@ -1,7 +1,7 @@
 import { Nullable } from 'typescript-nullable';
 import { Arguments } from 'yargs';
 import { RC } from '../../cli';
-import { State } from './types';
+import { State, Args } from './types';
 
 const getEnv = (env: unknown): string => {
   if (typeof env === 'string') {
@@ -12,7 +12,7 @@ const getEnv = (env: unknown): string => {
 };
 
 const initialState = (
-  args: Arguments,
+  args: Arguments<Args>,
   { rc, cwd }: { rc: Nullable<RC | Error>; cwd: string }
 ): State => ({
   rc,
